Move TransactionStatusHandler/useMessages/useEvents to own files
